Create linkedlist_implementation
This  contians a menu-driven code of doubly linked list ,its implementation and use of various functions like insert_at_head, insert_at_end, print, getlength. just one problem is there, when i am creating a memory to store head then data already sets to a garbage value, and this complicates the situation :,) as that garbage value or zero is always there in the linked list. If you can then please make the neccesary changes in the code.